requirements.txt
allure_python_commons==2.13.5
xlrd==1.2.0
pytest==6.2.5
selenium==3.141.0
Appium_Python_Client==1.0.2
docker_py==1.10.6
PyYAML==5.3.1
